Feature layout

+++ app/controllers/posts_controller.rb
@@ -3,13 +3,14 @@ class PostsController < ApplicationController
   before_action :set_categorys, only: [:new, :create, :edit, :update]
 
   def index
-    @posts = Post.all
+    @posts = Post.all.page(params[:page]).per(10)
     @categorys = Category.all
 
     if params[:id]
       @category = Category.find(params[:id])
-      @posts = @category.posts.order(created_at: :desc)
+      @posts = @category.posts.order(created_at: :desc).page(params[:page]).per(10)
     end
+
   end
 
   def show
@@ -17,25 +18,34 @@ class PostsController < ApplicationController
     @category = @post.category
     @user = @post.user
     @post_review = PostReview.new
-    @post_reviews = @post.post_reviews
+    @post_reviews = @post.post_reviews.page(params[:page]).per(10)
     @like = Like.new
-    new_history = @post.browsing_histories.new
-    new_history.user_id = current_user.id
-    if current_user.browsing_histories.exists?(post_id: "#{params[:id]}")
+
+    if user_signed_in?
+      new_history = @post.browsing_histories.new
+      new_history.user_id = current_user.id
+    end
+
+    if user_signed_in? && current_user.browsing_histories.exists?(post_id: "#{params[:id]}")
       old_history = current_user.browsing_histories.find_by(post_id: "#{params[:id]}")
       old_history.destroy
     end
-    new_history.save
-    histories_stock_limit = 10
-    histories = current_user.browsing_histories.all
-    if histories.count > histories_stock_limit
+
+    if user_signed_in?
+      new_history.save
+      histories_stock_limit = 10
+      histories = current_user.browsing_histories.all
+    end
+
+    if user_signed_in? && histories.count > histories_stock_limit
       histories[0].destroy
     end
+  rescue ActiveRecord::RecordNotFound
+    redirect_to posts_url, notice:"対象の投稿はありませんでした"
   end
 
   def search
-    @posts = Post.search(params[:search])
-    binding.pry
+    @posts = Post.search(params[:search]).page(params[:page]).per(10)
   end
 
   def new
@@ -45,11 +55,13 @@ class PostsController < ApplicationController
   def create
     @post = Post.new(post_params)
     @post.user_id = current_user.id
+
     if @post.save
       redirect_to post_path(@post.id), notice:"投稿しました"
     else
       render :new
     end
+
   end
 
   def edit
@@ -78,14 +90,6 @@ class PostsController < ApplicationController
 
   def set_categorys
     @categorys = Category.all
-  end
-
-  def self.search(search)
-    if search
-      Post.where(['content LIKE ?', "%#{search}%"])
-    else
-      Post.all
-    end
   end
 
 end
